% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamsteR.R
\name{check_loading_state}
\alias{check_loading_state}
\title{FUNCTION_TITLE}
\usage{
check_loading_state(module_file, check_state)
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
