% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamsteR-package.R
\docType{package}
\name{hamsteR-package}
\alias{hamsteR}
\alias{hamsteR-package}
\title{hamsteR: Module loading features}
\description{
The analysis of larger or more complex data sets usually takes place in many individual steps. Each of these steps uses input data, possibly producing intermediate products and finally end products. The end products then go back into subsequent steps. This is commonly called 'workflow'.
}
\keyword{internal}
