% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamsteR-package.R
\docType{package}
\name{hamsteR}
\alias{hamsteR}
\alias{hamsteR-package}
\title{hamsteR: A conditional \code{source} function}
\description{
The hamsteR package provides a conditional source function.
}
\details{
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

\keyword{internal}
