% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamsteR.R
\name{is_managed}
\alias{is_managed}
\alias{contains,}
\alias{managed}
\title{Check if the file is managed by \pkg{hamsteR}}
\usage{
is_managed(file)
}
\arguments{
\item{file}{file name of the script file to load}
}
\description{
checks if the given file is already managed by \pkg{hamsteR}
}
\examples{
\dontrun{
if(interactive()){
 set_loading_state("another_script.R", "undefined")
 is_managed("another_script.R") # TRUE
 reset()
 is_managed("another_script.R") # FALSE
}}
}
\seealso{
\code{\link{set_loading_state}} for setting the loading state of a file. This file is now \emph{managed} by \pkg{hamsteR}.
\code{\link{reset}} to delete all loading states.
}
