% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamsteR.R
\name{Load}
\alias{Load}
\title{FUNCTION_TITLE}
\usage{
Load(module_file, force_load = FALSE)
}
\arguments{
\item{module_file}{PARAM_DESCRIPTION}

\item{OnlyIfNotAlreadyLoaded}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
