% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamsteR.R
\name{loading_state_of}
\alias{loading_state_of}
\alias{get_loading_state}
\title{Determine the loading state of the given file}
\usage{
loading_state_of(file)
}
\arguments{
\item{file}{file name of the script file to load}
}
\value{
the state of the module, a value out of 'undefined', 'loading', 'loaded'
}
\description{
Returns the loading state of the given file
}
\details{
hamster is capable of managing loading state for file. With
this function you can determine the current loading state of the given
file. The loading state can be changed by \code{\link{set_loading_state}}
}
\examples{
\dontrun{
if(interactive()){
 ## check state 'loaded'
 set_loading_state(file="TEST01.R", state="loaded")
 loading_state_of("TEST01.R")  # "loaded"

 ## check state 'loading'
 set_loading_state(file="TEST01.R", state="loading")
 loading_state_of("TEST01.R")  # "loading"

 ## check state 'undefinded'
 set_loading_state(file="TEST01.R", state="undefined")
 loading_state_of("TEST01.R")  # "undefined"
}}
}
\seealso{
\code{\link{set_loading_state}} for setting the loading state, \code{\link{source_ifnotloaded}} for conditional loading and parsing a file
}
