% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamsteR.R
\name{reset}
\alias{reset}
\title{Clears all module states}
\usage{
reset()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Clears all module states so that all modules are reloaded the next time they are used
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
