% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamsteR.R
\name{set_all_unloaded}
\alias{set_all_unloaded}
\title{FUNCTION_TITLE}
\usage{
set_all_unloaded()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
