% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamsteR.R
\name{status}
\alias{status}
\title{Status of all managed files}
\usage{
status()
}
\value{
a data table, module names as row.names,
loading state in column 'loading_state'
}
\description{
returns a data.frame with the state of managed files
}
\examples{
\dontrun{
if(interactive()){
 set_loading_state("TEST_01.R", "loaded")
 set_loading_state("TEST_02.R", "undefined")
 status()
 ##           loading_state
 ## TEST_01.R loaded
 ## TEST_02.R undefined
}}
}
