% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamsteR.R
\name{reset}
\alias{reset}
\title{Clears all states}
\usage{
reset()
}
\description{
Clears all states so that all files simply will be loaded and parsed the next time
}
\examples{
\dontrun{
if(interactive()){
 source_ifnotloaded("TEST.R")
 set_loading_state("TEST.R", "loaded")
 source_ifnotloaded("TEST.R") # not loaded again because the loading state of the file was changed to 'loaded'
 is_managed("TEST.R") # TRUE
 reset() # clear all loading states
 is_managed("TEST.R") # FALSE
 source_ifnotloaded("TEST.R") # loaded again
 is_managed("TEST.R") # TRUE
 }
}
}
\seealso{
\code{\link{set_loading_state}} to set the loading state and \code{\link{loading_state_of}} to get the current loading state of a file
}
